refactor package namespaces (ref issue #15)
