+++ src/nz/cri/gns/NSHM/opensha/ruptures/downDipSubSectTest/DownDipTestRupSetBuilder.java
@@ -1,22 +1,12 @@
-package scratch.kevin.ucerf3.downDipSubSectTest.demo;
+package nz.cri.gns.NSHM.opensha.ruptures.downDipSubSectTest;
 
-import java.io.BufferedInputStream;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.io.InputStream;
 import java.util.ArrayList;
-import java.util.Collection;
-//import java.util.HashMap;
+import java.util.HashMap;
 import java.util.List;
-//import java.util.Map;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.util.Map;
 
-import javax.imageio.stream.FileImageInputStream;
-
-import org.dom4j.DocumentException;
 import org.opensha.commons.geo.Location;
 import org.opensha.commons.util.FaultUtils;
 import org.opensha.commons.util.IDPairing;
@@ -32,85 +22,72 @@ import org.opensha.sha.faultSurface.FaultSection;
 import org.opensha.sha.faultSurface.FaultTrace;
 import org.opensha.sha.faultSurface.SimpleFaultData;
 
-import com.google.common.base.Joiner;
 import com.google.common.base.Preconditions;
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Lists;
 
 import scratch.UCERF3.FaultSystemRupSet;
-import scratch.UCERF3.enumTreeBranches.FaultModels;
 import scratch.UCERF3.enumTreeBranches.ScalingRelationships;
-
+import scratch.UCERF3.inversion.SectionCluster;
+import scratch.UCERF3.inversion.SectionClusterList;
+import scratch.UCERF3.inversion.SectionConnectionStrategy;
+import scratch.UCERF3.utils.DeformationModelFetcher;
 import scratch.UCERF3.utils.FaultSystemIO;
-import scratch.kevin.ucerf3.downDipSubSectTest.DownDipSubSectBuilder;
-import scratch.kevin.ucerf3.downDipSubSectTest.DownDipTestPermutationStrategy;
-import scratch.kevin.ucerf3.downDipSubSectTest.RectangularityFilter;
 
-public class HikurangiCrustalDemoTwo {
+public class DownDipTestRupSetBuilder {
 
-	static DownDipSubSectBuilder downDipBuilder;
-	
-	public static void main(String[] args) throws DocumentException, IOException {
-		// maximum sub section length (in units of DDW)
-		double maxSubSectionLength = 0.5;
-		// max distance for linking multi fault ruptures, km
-		//double maxDistance = 0.5d;
+	public static void main(String[] args) throws IOException {
+		File outputFile = new File("/tmp/down_dip_sub_sect_rup_set.zip");
 		
-		File outputFile = new File("/tmp/rupSetLowerNIAndInterface30km.zip");
-	
-		File fsdFile = new File("./data/FaultModels/cfm_test.xml");
+		// we're going to manually build faults here
+		// first, build a big fault with down-dip subsections
+		String sectName = "Test SubSect Down-Dip Fault";
+		int sectID = 0;
+		int startID = 0;
+		double upperDepth = 0d;
+		double lowerDepth = 30d;
+		double dip = 20d;
+//		int numDownDip = 5;
+//		int numAlongStrike = 6;
+		int numDownDip = 20;
+		int numAlongStrike = 30;
+		FaultTrace trace = new FaultTrace(sectName);
+		trace.add(new Location(34, -119, upperDepth));
+		trace.add(new Location(34.1, -118.75, upperDepth));
+		trace.add(new Location(34.15, -118.5, upperDepth));
+		trace.add(new Location(34.1, -118.25, upperDepth));
+		trace.add(new Location(34, -118, upperDepth));
 		
-		// load in the fault section data ("parent sections")
-		List<FaultSection> fsd = FaultModels.loadStoredFaultSections(fsdFile);
+		SimpleFaultData faultData = new SimpleFaultData(dip, lowerDepth, upperDepth, trace);
+		double aveRake = 90d;
 		
-		Stream<Integer> myStream01 = Stream.of(83, 84, 85, 86, 87);
-		Collection<Integer> sectsWairarapa = myStream01.collect(Collectors.toCollection(ArrayList::new));
-
-		Stream<Integer> myStream02 = Stream.of(89, 90, 91, 92, 93);
-		Collection<Integer> sectsWellington = myStream02.collect(Collectors.toCollection(ArrayList::new));
+		DownDipSubSectBuilder downDipBuilder = new DownDipSubSectBuilder(sectName, sectID, startID,
+				faultData, aveRake, numAlongStrike, numDownDip);
 		
-		List<Integer> sectsToKeep = Lists.newArrayList(Iterables.concat(sectsWairarapa, sectsWellington));
-
-   	    if (sectsToKeep != null && !sectsToKeep.isEmpty()) {
-		 	System.out.println("Only keeping these parent fault sections: "
-		 			+Joiner.on(",").join(sectsToKeep));
-		 	// iterate backwards as we will be removing from the list
-		 	for (int i=fsd.size(); --i>=0;)
-		 		if (!sectsToKeep.contains(fsd.get(i).getSectionId()))
-		 			fsd.remove(i);
-		 }
-			
-		
-		// build the subsections
 		List<FaultSection> subSections = new ArrayList<>();
-		int sectIndex = 0;
-		for (FaultSection parentSect : fsd) {
-			double ddw = parentSect.getOrigDownDipWidth();
-			double maxSectLength = ddw*maxSubSectionLength;
-			// the "2" here sets a minimum number of sub sections
-			List<? extends FaultSection> newSubSects = parentSect.getSubSectionsList(maxSectLength, sectIndex, 2);
-			subSections.addAll(newSubSects);
-			sectIndex += newSubSects.size();
-		}
-		
-		System.out.println(subSections.size()+" Sub Sections");
-		
-		String sectName = "Interface Fault 30km";
-//		int sectID = 0;
-		int startID = subSections.size();
-		
-		FaultSection interfaceParentSection = new FaultSectionPrefData();
-		interfaceParentSection.setSectionId(10000);
-		interfaceParentSection.setSectionName(sectName);
-				
-		File initialFile = new File("./data/FaultModels/subduction_tile_parameters_30.csv");
-	    InputStream inputStream = new FileInputStream(initialFile);
-		downDipBuilder = new DownDipSubSectBuilder(sectName, interfaceParentSection, startID, inputStream);
-		
-		// Add the interface subsections
 		subSections.addAll(downDipBuilder.getSubSectsList());
+		System.out.println("Have "+subSections.size()+" sub-sections for "+sectName);
+		startID = subSections.size();
 		
-		System.out.println("Have "+subSections.size()+" sub-sections in total");		
+		// now lets add a nearby crustal fault that it can interact with
+		sectName = "Test Crustal Fault";
+		sectID = 1;
+		upperDepth = 0d;
+		lowerDepth = 14d;
+		dip = 90d;
+		trace = new FaultTrace(sectName);
+		trace.add(new Location(33.5, -117.5, upperDepth));
+		trace.add(new Location(34.1, -118.25, upperDepth));
+		FaultSectionPrefData crustalSect = new FaultSectionPrefData();
+		crustalSect.setSectionId(sectID);
+		crustalSect.setSectionName(sectName);
+		crustalSect.setFaultTrace(trace);
+		crustalSect.setAveUpperDepth(upperDepth);
+		crustalSect.setAveLowerDepth(lowerDepth);
+		crustalSect.setAseismicSlipFactor(0d);
+		crustalSect.setAveDip(dip);
+		double maxSectLength = 0.5*crustalSect.getOrigDownDipWidth();
+		
+		subSections.addAll(crustalSect.getSubSectionsList(maxSectLength, startID, 2));
+		System.out.println("Have "+subSections.size()+" sub-sections in total");
 		
 		for (int s=0; s<subSections.size(); s++)
 			Preconditions.checkState(subSections.get(s).getSectionId() == s,
@@ -118,6 +95,7 @@ public class HikurangiCrustalDemoTwo {
 		
 		// instantiate plausibility filters
 		List<PlausibilityFilter> filters = new ArrayList<>();
+		
 		int minDimension = 1; // minimum numer of rows or columns
 		double maxAspectRatio = 5d; // max aspect ratio of rows/cols or cols/rows
 		filters.add(new RectangularityFilter(downDipBuilder, minDimension, maxAspectRatio));
@@ -183,10 +161,12 @@ public class HikurangiCrustalDemoTwo {
 		}
 		
 		String info = "Test down-dip subsectioning rup set";
+		
 		FaultSystemRupSet rupSet = new FaultSystemRupSet(subSections, sectSlipRates, null, sectAreasReduced,
 				rupsIDsList, rupMags, rupRakes, rupAreas, rupLengths, info);
 		FaultSystemIO.writeRupSet(rupSet, outputFile);
 		
 		System.out.println("All done!"); 
 	}
+
 }
