+++ src/nz/cri/gns/NSHM/opensha/ruptures/downDipSubSectTest/RectangularityFilter.java
@@ -1,4 +1,4 @@
-package scratch.kevin.ucerf3.downDipSubSectTest;
+package nz.cri.gns.NSHM.opensha.ruptures.downDipSubSectTest;
 
 import org.opensha.sha.earthquake.faultSysSolution.ruptures.ClusterRupture;
 import org.opensha.sha.earthquake.faultSysSolution.ruptures.FaultSubsectionCluster;
